backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pymongo==4.5.0
pydantic==2.3.0
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
openai==0.28.0